+++ packages/api/src/index.ts
@@ -31,7 +31,11 @@ export {
   type SendBoxToolbarMiddlewareRequest
 } from './middleware/SendBoxToolbarMiddleware';
 export { default as normalizeStyleOptions } from './normalizeStyleOptions';
-export { type StrictStyleOptions, type default as StyleOptions } from './StyleOptions';
+export {
+  type RenderFeedbackFormOverrideComponentContext,
+  type StrictStyleOptions,
+  type default as StyleOptions
+} from './StyleOptions';
 export { type ActivityStatusMiddleware, type RenderActivityStatus } from './types/ActivityStatusMiddleware';
 export {
   type AttachmentForScreenReaderComponentFactory,